+++ lib/vm/src/errors.rs
@@ -55,10 +55,10 @@ pub enum Error {
         #[cfg(feature = "backtraces")]
         backtrace: snafu::Backtrace,
     },
-    // Similar to RuntimeErr but with custom message
-    #[snafu(display("Error during wasm execution: {}", msg))]
-    ExecutionErr {
-        msg: &'static str,
+    #[snafu(display("Region too small. Got {}, required {}", size, required))]
+    RegionTooSmallErr {
+        size: usize,
+        required: usize,
         #[cfg(feature = "backtraces")]
         backtrace: snafu::Backtrace,
     },